Split ControlBar components into several sub components
